Do not set self_bitmap flag for QXL_DRAW_COPY command
Setting this flag on this command cause some performance issue.
When the server see this flag it updates the frame buffer and copy
part of it in a new allocated image. However for this command the
image is not used wasting only CPU and memory.
Also the frame buffer update causes the elimination optimisation
to be less effective causing more network usage.

Signed-off-by: Frediano Ziglio <fziglio@redhat.com>
Acked-by: Yuri Benditovich <yuri.benditovich@daynix.com>